+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Respect group `showDecimals` setting in the shared balance image card ("WHO OWES WHOM" and "EXPENSE SHARES").
+
 ## [1.11.0] - 2026-09-11
 
 ### Added